初始化数据可视化Koa平台
