--- portrait/src/com/jiavan/weibo/crawler/test/ConnectionFactoryTest.java
@@ -1,24 +1,0 @@
-package com.jiavan.weibo.crawler.test;
-
-
-import com.jiavan.weibo.crawler.jdbc.ConnectionFactory;
-
-import java.sql.Connection;
-
-/**
- * Connection Factory Singleton unit test
- * Created by Jiavan on 2017/4/5.
- */
-public class ConnectionFactoryTest {
-    public static void main(String[] args) {
-        try {
-            ConnectionFactory cf = ConnectionFactory.getInstance();
-            Connection connection = cf.getConnection();
-
-            System.out.print(connection);
-            connection.close();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-}
