+++ web/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "weibo-portrait-web",
+  "version": "0.1.0",
+  "description": "A simple weibo user portrait data visualization web platform",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+ssh://git@github.com/13391/wb.git"
+  },
+  "keywords": [
+    "data",
+    "visualization"
+  ],
+  "author": "Jiavan, jiavan.com@gmail.com",
+  "license": "GPL-3.0",
+  "bugs": {
+    "url": "https://github.com/13391/wb/issues"
+  },
+  "homepage": "https://github.com/13391/wb#readme",
+  "dependencies": {
+    "koa-static": "^3.0.0",
+    "mysql": "^2.13.0"
+  }
+}
