+++ web/static/script/index.html
@@ -1,0 +1,1 @@
+hello
